requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
aiohappyeyeballs==2.4.3
    # via aiohttp
aiohttp==3.10.10
    # via gcsfs
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.6.2
    # via
    #   httpx
    #   openai
async-timeout==4.0.3
    # via aiohttp
attrs==24.2.0
    # via aiohttp
bottleneck==1.4.1
    # via pandas
cachetools==5.5.0
    # via google-auth
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.0
    # via requests
cryptography==43.0.1
    # via unstructured-client
db-dtypes==1.3.0
    # via pandas-gbq
decorator==5.1.1
    # via gcsfs
deprecated==1.2.14
    # via pikepdf
distro==1.9.0
    # via openai
eval-type-backport==0.2.0
    # via unstructured-client
exceptiongroup==1.2.2
    # via anyio
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
fsspec==2024.9.0
    # via gcsfs
gcsfs==2024.9.0.post1
    # via pandas
google-api-core[grpc]==2.21.0
    # via
    #   -r requirements.in
    #   google-cloud-bigquery
    #   google-cloud-core
    #   google-cloud-documentai
    #   google-cloud-documentai-toolbox
    #   google-cloud-storage
    #   google-cloud-vision
    #   pandas-gbq
google-auth==2.35.0
    # via
    #   gcsfs
    #   google-api-core
    #   google-auth-oauthlib
    #   google-cloud-bigquery
    #   google-cloud-core
    #   google-cloud-documentai
    #   google-cloud-storage
    #   google-cloud-vision
    #   pandas-gbq
    #   pydata-google-auth
google-auth-oauthlib==1.2.1
    # via
    #   gcsfs
    #   pandas-gbq
    #   pydata-google-auth
google-cloud-bigquery==3.26.0
    # via
    #   google-cloud-documentai-toolbox
    #   pandas-gbq
google-cloud-core==2.4.1
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
google-cloud-documentai==2.33.0
    # via
    #   -r requirements.in
    #   google-cloud-documentai-toolbox
google-cloud-documentai-toolbox==0.14.0a0
    # via -r requirements.in
google-cloud-storage==2.18.2
    # via
    #   -r requirements.in
    #   gcsfs
    #   google-cloud-documentai-toolbox
google-cloud-vision==3.7.4
    # via google-cloud-documentai-toolbox
google-crc32c==1.6.0
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.7.2
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
googleapis-common-protos[grpc]==1.65.0
    # via
    #   google-api-core
    #   grpc-google-iam-v1
    #   grpcio-status
grpc-google-iam-v1==0.13.1
    # via google-cloud-documentai-toolbox
grpcio==1.66.2
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.66.2
    # via google-api-core
h11==0.14.0
    # via httpcore
httpcore==1.0.6
    # via httpx
httpx==0.27.2
    # via
    #   openai
    #   unstructured-client
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
immutabledict==4.2.0
    # via google-cloud-documentai-toolbox
intervaltree==3.1.0
    # via google-cloud-documentai-toolbox
jinja2==3.1.4
    # via google-cloud-documentai-toolbox
jiter==0.6.1
    # via openai
jsonpath-python==1.0.6
    # via unstructured-client
llvmlite==0.43.0
    # via numba
lxml==5.3.0
    # via pikepdf
markupsafe==3.0.1
    # via jinja2
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via typing-inspect
nest-asyncio==1.6.0
    # via unstructured-client
numba==0.60.0
    # via pandas
numexpr==2.10.1
    # via pandas
numpy==1.26.4
    # via
    #   bottleneck
    #   db-dtypes
    #   google-cloud-documentai-toolbox
    #   numba
    #   numexpr
    #   pandas
    #   pandas-gbq
    #   pyarrow
oauthlib==3.2.2
    # via requests-oauthlib
openai==1.51.2
    # via -r requirements.in
packaging==24.1
    # via
    #   db-dtypes
    #   google-cloud-bigquery
    #   pandas-gbq
    #   pikepdf
pandas[gcp,performance]==2.2.3
    # via
    #   db-dtypes
    #   google-cloud-documentai-toolbox
    #   pandas-gbq
pandas-gbq==0.23.2
    # via pandas
pikepdf==8.15.1
    # via google-cloud-documentai-toolbox
pillow==10.4.0
    # via
    #   google-cloud-documentai-toolbox
    #   pikepdf
propcache==0.2.0
    # via yarl
proto-plus==1.24.0
    # via
    #   google-api-core
    #   google-cloud-documentai
    #   google-cloud-documentai-toolbox
    #   google-cloud-vision
protobuf==5.28.2
    # via
    #   google-api-core
    #   google-cloud-documentai
    #   google-cloud-vision
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   proto-plus
pyarrow==15.0.2
    # via
    #   db-dtypes
    #   google-cloud-documentai-toolbox
    #   pandas-gbq
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.1
    # via google-auth
pycparser==2.22
    # via cffi
pydantic==2.9.2
    # via
    #   openai
    #   unstructured-client
pydantic-core==2.23.4
    # via pydantic
pydata-google-auth==1.8.2
    # via pandas-gbq
pypdf==5.0.1
    # via unstructured-client
python-dateutil==2.8.2
    # via
    #   google-cloud-bigquery
    #   pandas
    #   unstructured-client
python-dotenv==1.0.1
    # via -r requirements.in
pytz==2024.2
    # via pandas
requests==2.32.3
    # via
    #   gcsfs
    #   google-api-core
    #   google-cloud-bigquery
    #   google-cloud-storage
    #   requests-oauthlib
    #   requests-toolbelt
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
requests-toolbelt==1.0.0
    # via unstructured-client
rsa==4.9
    # via google-auth
six==1.16.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
sortedcontainers==2.4.0
    # via intervaltree
tabulate==0.9.0
    # via google-cloud-documentai-toolbox
tqdm==4.66.5
    # via openai
typing-extensions==4.12.2
    # via
    #   anyio
    #   multidict
    #   openai
    #   pydantic
    #   pydantic-core
    #   pypdf
    #   typing-inspect
typing-inspect==0.9.0
    # via unstructured-client
tzdata==2024.2
    # via pandas
unstructured-client==0.26.1
    # via -r requirements.in
urllib3==2.2.3
    # via requests
wrapt==1.16.0
    # via deprecated
yarl==1.15.2
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
